refactor(admin): extract useInboxCount hook; fix order and tab badge
